% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlesky.R
\name{boot}
\alias{boot}
\title{Parametric bootstrap for Ne and regression coefficients from a list of trees or a multiPhylo object.}
\usage{
boot(fit, trees, ncpu = 1)
}
\arguments{
\item{fit}{mlesky fitted object (mlskygrid class)}

\item{trees}{list of trees or multiTree object}

\item{ncpu}{Number of CPUs}
}
\value{
A list of fitted mlesky models (length equal to the number of trees provided) with confidence intervals for Ne and regression coefficients
}
\description{
This will receive a mlesky fit and coalescent trees as parameters and calculate confidence intervals based on these.
}
