% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{ddSimCoal}
\alias{ddSimCoal}
\title{Simulation of coalescent dated phylogeny using deterministic/'sideways' distribution of coalescent times}
\usage{
ddSimCoal(
  dates = 1990:2010,
  alphaFun = function(x) {     return(10) },
  alphaMin = NA,
  guessRootTime = NA,
  res = 1000,
  A_tol = 0.01
)
}
\arguments{
\item{dates}{Sampling dates}

\item{alphaFun}{Population size function Ne(t)}

\item{guessRootTime}{Optional (but recommended) guess for when root may be reached.}

\item{res}{Time axis resolution}

\item{A_tol}{Tolerance for reaching time of root >0}
}
\value{
A simulated dated phylogeny
}
\description{
Simulation of coalescent dated phylogeny using deterministic/'sideways' distribution of coalescent times
}
