% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{mlesky}
\alias{mlesky}
\title{mlesky: Maximum likelihood estimation of effective population size through time, growth rates, and phylodynamic regressions}
\description{
mlesky is a R package to perform Maximum likelihood estimation of effective population size through time, growth rates, and phylodynamic regressions.
}
\seealso{
https://github.com/emvolz-phylodynamics/mlesky
}
\author{
Erik Volz \email{erik.volz@gmail.com}
}
