% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlskygrid.R
\name{mlskygrid}
\alias{mlskygrid}
\title{Maximum likelihood non-parametric estimation of effective population size through time}
\usage{
mlskygrid(tre, res = 25, tau = 1, tau_lower = NULL,
  tau_upper = NULL, tau_tol = 0.001, ncross = 5, ncpu = 1,
  quiet = FALSE, NeStartTimeBeforePresent = Inf, ...)
}
\arguments{
\item{tre}{A dated phylogeny in ape::phylo format (see documentation for ape)}

\item{res}{Length of time axis over which to estimate Ne(t) (integer)}

\item{tau}{Precision parameter. Larger values generate smoother trajectories of Ne(t). If NULL, will optimize using cross-validation.}

\item{tau_lower}{Lower bound for precision parameter if estimating}

\item{tau_upper}{Upper bound for precision parameter if estimating}

\item{tau_tol}{Optimization tolerance when optimizing tau by cross-validation}

\item{ncross}{Number of folds in cross-validation}

\item{ncpu}{If doing cross-validation, each fold will be handled in parallel if ncpu > 1 (see parallel package)}

\item{quiet}{Provide verbose output from optimizer?}

\item{NeStartTimeBeforePresent}{If <Inf, will only estimate Ne(t) between the most recent sample and this time before the most recent sample}
}
\value{
A fitted model including effective size through time
}
\description{
Maximum likelihood non-parametric estimation of effective population size through time
}
