% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlesky.R
\name{optim_res_aic}
\alias{optim_res_aic}
\title{Optimize the skygrid time axis resolution using AIC criterion}
\usage{
optim_res_aic(tree, res = c(1, 3, 5, seq(10, 100, by = 10)), ncpu = 1, ...)
}
\arguments{
\item{tree}{A dated phylogeny in ape::phylo format}

\item{res}{A vector of time axis resolution parameters to test}

\item{ncpu}{Integer number of cores to use with parallel processing}

\item{...}{Remaining parameters are passed to mlskygrid}
}
\description{
Optimize the skygrid time axis resolution using AIC criterion
}
