% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlesky.R
\name{parboot}
\alias{parboot}
\title{Parametric bootstrap for Ne and regression coefficients.}
\usage{
parboot(fit, nrep = 200, ncpu = 1)
}
\arguments{
\item{fit}{mlesky fit}

\item{nrep}{Number of simulations}
}
\value{
A fitted mlesky model with updated confidence intervals for Ne and regression coefficients
}
\description{
This will simulate coalescent trees conditional on the provided mlesky estimate of Ne. 
The model is refitted to each coalescent tree to provide an estimate of the standard error of Ne estimates over time.
}
