% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlesky.R
\name{plot.mlskygrid}
\alias{plot.mlskygrid}
\title{Plot mlskygrid}
\usage{
\method{plot}{mlskygrid}(x, growth = FALSE, ggplot = FALSE, logy, ...)
}
\arguments{
\item{x}{A fitted object}

\item{growth}{If TRUE will plot estimated growth rate instead of Ne(t)}

\item{ggplot}{If TRUE, returns a ggplot2 figure}

\item{logy}{If TRUE, the plot is returned with logarithmic y-axis (default is TRUE for Ne plot and FALSE for growth plot)}

\item{...}{Additional parameters are passed to ggplot or the base plotting function}
}
\value{
Plotted object
}
\description{
Plot mlskygrid
}
