% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{plotBoth}
\alias{plotBoth}
\title{Plot a dated phylogeny and alpha function}
\usage{
plotBoth(tree, alphaFun)
}
\arguments{
\item{tree}{Dated phylogeny}

\item{alphaFun}{Population size function Ne(t)}
}
\value{
Figure
}
\description{
Plot a dated phylogeny and alpha function
}
