% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simCoal}
\alias{simCoal}
\title{Simulation of coalescent dated phylogeny}
\usage{
simCoal(
  dates = 1990:2010,
  alphaFun = function(x) {
     return(10)
 },
  alphaMin = NA
)
}
\arguments{
\item{dates}{Sampling dates}

\item{alphaFun}{Population size function Ne(t)}

\item{alphaMin}{Minimum value of alphaFun}
}
\value{
A simulated dated phylogeny
}
\description{
Simulation of coalescent dated phylogeny
}
