% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlesky.R
\name{suggest_res}
\alias{suggest_res}
\title{Heuristic selection of grid resolution. Selects res based on  
reduction in MSE of coalescent times relative to nearest grid point.}
\usage{
suggest_res(tree, th = 0.001)
}
\arguments{
\item{tree}{A dated phylogeny in ape::phylo format}

\item{th}{Precision parameter}
}
\description{
Heuristic selection of grid resolution. Selects res based on  
reduction in MSE of coalescent times relative to nearest grid point.
}
