% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlesky.R
\name{mlskygrid}
\alias{mlskygrid}
\title{Maximum likelihood non-parametric estimation of effective population size through time}
\usage{
mlskygrid(
  tre,
  sampleTimes = NULL,
  res = 25,
  tau = 1,
  tau_lower = NULL,
  tau_upper = NULL,
  tau_tol = 0.001,
  ncross = 5,
  ncpu = 1,
  quiet = TRUE,
  NeStartTimeBeforePresent = Inf,
  ne0 = NULL,
  adapt_time_axis = FALSE,
  formula = NULL,
  formula_order = c(0, 1, 2),
  data = NULL
)
}
\arguments{
\item{tre}{A dated phylogeny in ape::phylo or treedater format (see documentation for ape). This can also be a multiPhylo or list of trees, in which case each is treated as a clade sampled from within the same population. In this case the sampleTimes vector should be supplied so that clades can be aligned in time.}

\item{sampleTimes}{An optional named vector of sample times for each taxon. Names should correspond to tip labels in trees. This is required if providing a list of trees.}

\item{res}{Length of time axis over which to estimate Ne(t) (integer). If NULL, will search for a good value (see *optim_res_aic*)}

\item{tau}{Precision parameter. Larger values generate smoother trajectories of Ne(t). If NULL, will optimize using cross-validation.}

\item{tau_lower}{Lower bound for precision parameter if estimating}

\item{tau_upper}{Upper bound for precision parameter if estimating}

\item{tau_tol}{Optimization tolerance when optimizing tau by cross-validation}

\item{ncross}{Number of folds in cross-validation}

\item{ncpu}{If doing cross-validation, each fold will be handled in parallel if ncpu > 1 (see parallel package)}

\item{quiet}{Provide verbose output from optimizer?}

\item{NeStartTimeBeforePresent}{If <Inf, will only estimate Ne(t) between the most recent sample and this time before the most recent sample}

\item{ne0}{Vector of length *res* giving starting conditions of Ne(t) for optimization}

\item{adapt_time_axis}{If TRUE will choose Ne(t) change points in periods with high frequency of phylogenetic branching}
}
\value{
A fitted model including effective size through time
}
\description{
Maximum likelihood non-parametric estimation of effective population size through time
}
